+++ CHANGELOG.rst
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+* Improved error-handling in MarketWatch.
 * Fixed a bug in ``dataset.update()`` not being able to handle duplicate l18 values.
 * Updated dataset.